+++ frontend/src/App.vue
@@ -6,12 +6,13 @@ import { createMap } from "./map.js";
 import AgencyServiceForm from "./components/AgencyServiceForm.vue";
 import FeedSummary from "./components/FeedSummary.vue";
 import ModeBar from "./components/ModeBar.vue";
-import NoticesPanel from "./components/NoticesPanel.vue";
 import RouteForm from "./components/RouteForm.vue";
 import SaveBar from "./components/SaveBar.vue";
 import StopInspector from "./components/StopInspector.vue";
+import TabBar from "./components/TabBar.vue";
 import TimetablePanel from "./components/TimetablePanel.vue";
 import TripForm from "./components/TripForm.vue";
+import ValidationReport from "./components/ValidationReport.vue";
 
 onMounted(createMap);
 </script>
@@ -19,15 +20,21 @@ onMounted(createMap);
 <template>
   <div id="sidebar">
     <h1>transitio</h1>
-    <FeedSummary />
-    <ModeBar />
-    <StopInspector />
-    <RouteForm />
-    <TripForm />
-    <AgencyServiceForm />
-    <TimetablePanel />
-    <NoticesPanel />
-    <SaveBar />
+    <TabBar />
+
+    <div v-show="store.activeTab === 'edit'">
+      <FeedSummary />
+      <ModeBar />
+      <StopInspector />
+      <RouteForm />
+      <TripForm />
+      <AgencyServiceForm />
+      <TimetablePanel />
+      <SaveBar />
+    </div>
+
+    <ValidationReport v-show="store.activeTab === 'report'" />
+
     <div id="status">{{ store.status }}</div>
   </div>
   <div id="map"></div>